author karma now displayed
