+++ src/components/story/Story.scss
@@ -1,11 +1,6 @@
 @import '/src/utils/shared-styles.scss';
 
 .story {
-  &__title-container {
-    align-self: flex-end;
-    width: 50%;
-  }
-
   &__title {
     text-decoration: none;
     color: black;
@@ -19,17 +14,14 @@
   &__url {
     color: cadetblue;
     margin-top: 0.3rem;
-  }
-
-  &__score {
-    display: flex;
-    justify-content: flex-end;
-    flex-grow: 2;
+    // width: 100%;
   }
 
   &__info {
     margin-top: 1rem;
     color: grey;
     font-style: italic;
+    flex-grow: 2;
+    align-items: flex-end;
   }
 }
