+++ src/components/story/Story.tsx
@@ -1,10 +1,13 @@
 import React, { useState, useEffect } from 'react';
 
-import { getStory } from '../../services/hnApi';
-import { getUser } from '../../services/userApi';
-import { selectStoryScore } from '../../utils/selectFields';
+// Api
+import { getStory } from '../../services/storyApi';
 
+// Asssets
 import image from './../../utils/assets/dummy-image-one.jpg';
+
+// Components
+import { User } from '../user-information/User';
 
 import './Story.scss';
 
@@ -12,43 +15,22 @@ export interface StoryItem {
   title: string;
   id: number;
   score: number;
-  time: any; // Date??
+  time: any;
   url: string;
-  by: any; //string
-}
-
-export interface UserItem {
-  karma: number;
-  id: string;
+  by: any;
 }
 
 interface StoryProps {
   storyItem?: StoryItem;
-  userItem?: UserItem;
   storyId: number;
-  // userId?: string;
-  // missing author karma right now
 }
 
-export const Story = ({ storyId, storyItem, userItem }: StoryProps) => {
+export const Story = ({ storyId, storyItem }: StoryProps) => {
   const [story, setStory] = useState({ ...storyItem });
-  const [user, setUser] = useState({ ...userItem });
-
-  // console.log('user', user);
 
   useEffect(() => {
     getStory(storyId).then(data => data && data.url && setStory(data));
-    // getUser(userId).then(userData => userData && setUser(userData));
   }, []);
-
-  // Random way to fix it but would it be possible to have it in another component where it's only the userId and then pass it in here as part of the return?? would probably need to map something
-
-  // useEffect(() => {
-  //   // getUser(story.by).then(data => console.log(data));
-  // }, []);
-
-  // const userId: string = story.by;
-  // console.log('story', story);
 
   const date = new Date(story.time * 1000).toLocaleDateString('en-GB', { hour: 'numeric', minute: 'numeric' });
 
@@ -56,7 +38,7 @@ export const Story = ({ storyId, storyItem, userItem }: StoryProps) => {
     <div className='page-container card'>
       <div className='direction-row'>
         <img className='story__image' src={image} alt='dummy-image' />
-        <div className='direction-column story__title-container'>
+        <div className='direction-column story__container'>
           <a href={story.url} className='story__title'>
             <h2>{story.title}</h2>
           </a>
@@ -65,12 +47,10 @@ export const Story = ({ storyId, storyItem, userItem }: StoryProps) => {
             {story.url}
           </a>
           <div className='direction-row story__info'>
-            <p>Posted by: {story.by}</p>
             <p>{date}</p>
-            <p>Karma: {'authorKarma'}</p>
+            <User userId={story.by} />
           </div>
         </div>
-        {/* <p className='story__score'>Story score: {story.score}</p> */}
       </div>
     </div>
   ) : null;
